+++ lib/src/data/data_source/local/db_adapter/account_db_adapter.dart
@@ -5,5 +5,5 @@ abstract class AccountDbAdapter {
 
   AccountHiveEntity? getAccountEntity(String userId);
 
-  Stream<AccountHiveEntity> listenAccountEntity();
+  Stream<AccountHiveEntity?> listenAccountEntity(String userId);
 }