+++ lib/src/data/repo_impl/post_repo_impl.dart
@@ -63,7 +63,14 @@ class PostRepoImpl extends PostRepo {
   @override
   Future<bool> flagPost(String postId) async {
     final data = await publicPostApiInterface.flagPost(postId);
-    return data;
+    await data.saveToDb<AmityPost>(dbAdapterRepo);
+
+    // Update the isFlaggedByMe
+    final post = dbAdapterRepo.postDbAdapter.getPostEntity(postId);
+    post.flaggedByMe = true;
+    post.save();
+
+    return true;
   }
 
   @override
@@ -75,7 +82,14 @@ class PostRepoImpl extends PostRepo {
   @override
   Future<bool> unflagPost(String postId) async {
     final data = await publicPostApiInterface.unflagPost(postId);
-    return data;
+    await data.saveToDb<AmityPost>(dbAdapterRepo);
+
+    // Update the isFlaggedByMe
+    final post = dbAdapterRepo.postDbAdapter.getPostEntity(postId);
+    post.flaggedByMe = false;
+    post.save();
+
+    return true;
   }
 
   @override