+++ lib/src/domain/repo/account_repo.dart
@@ -2,5 +2,5 @@ import 'package:amity_sdk/src/data/data_source/data_source.dart';
 
 abstract class AccountRepo {
   AccountHiveEntity? getAccount(String userId);
-  Stream<AccountHiveEntity> listenAccount(String userId);
+  Stream<AccountHiveEntity?> listenAccount(String userId);
 }