+++ lib/src/public/model_extenstion/amity_post_extension.dart
@@ -80,4 +80,15 @@ extension AmityPostExtension on AmityPost {
 
     return controller.stream;
   }
+
+  /// check if post is flagged by me
+  bool get isFlaggedByMe {
+    if (hashFlag == null) return false;
+    return (flaggedByMe ?? false) ||
+        BloomFilter(
+                hash: (hashFlag!['hash'] as String),
+                m: hashFlag!['bits'] as int,
+                k: hashFlag!['hashes'] as int)
+            .mightContains(AmityCoreClient.getUserId());
+  }
 }