+++ lib/src/public/query_builder/comment/comment_flag_query_builder.dart
@@ -1,9 +1,12 @@
 import 'package:amity_sdk/src/domain/domain.dart';
 
+/// Comment Repor Query Builder
 class CommentFlagQueryBuilder {
   final CommentFlagUsecase _commentFlagUsecase;
   final CommentUnflagUsecase _commentUnflagUsecase;
   final String _commentId;
+
+  /// Init [CommentFlagQueryBuilder]
   CommentFlagQueryBuilder(
       {required CommentFlagUsecase commentFlagUsecase,
       required CommentUnflagUsecase commentUnflagUsecase,
@@ -11,10 +14,13 @@ class CommentFlagQueryBuilder {
       : _commentFlagUsecase = commentFlagUsecase,
         _commentUnflagUsecase = commentUnflagUsecase,
         _commentId = commentId;
+
+  /// Flag Comment
   Future<bool> flag() {
     return _commentFlagUsecase.get(_commentId);
   }
 
+  /// Unflag Comment
   Future<bool> unflag() {
     return _commentUnflagUsecase.get(_commentId);
   }